Tests/LibWeb/Text/expected/input-delete.txt

--- a ---
Before: foo👩🏼‍❤️‍👨🏻bar
After: foobar
--- b ---
Before: foo👩🏼‍❤️‍👨🏻bar
After: foobar
--- c ---
Before: foo bar
After: foo 
--- d ---
Before: foo bar
After:  bar
